openmemory/api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.5
fastapi-pagination>=0.12.0
mem0ai>=0.1.92,<0.2.0
mcp[cli]>=1.3.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0.0
tenacity==9.1.2
supabase>=2.5.0,<3.0.0
feedparser>=6.0.0
python-dateutil>=2.8.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
psutil>=5.9.0
starlette==0.38.2
sse-starlette==2.1.3
posthog>=3.0.0
email-validator>=2.0.0
stripe>=12.0.0
